pyproject.toml
[tool.poetry]
name = "sentinex"
version = "0.1.0"
description = "AI Operating System for mid-market companies"
authors = ["Sentinex"]
readme = "README.md"
license = "Proprietary"
package-mode = false

[tool.poetry.dependencies]
python = "^3.12"
django = "^5.1"
django-tenants = "^3.7"
djangorestframework = "^3.15"
django-allauth = "^64.0"
django-cryptography-django5 = "^2.2"
django-environ = "^0.11"
django-celery-beat = "^2.7"
celery = { version = "^5.4", extras = ["redis"] }
redis = "^5.1"
psycopg = { version = "^3.2", extras = ["binary", "pool"] }
pgvector = "^0.3"
anthropic = "^0.40"
openai = "^1.60"
pydantic = "^2.9"
pyyaml = "^6.0"
tenacity = "^9.0"
weasyprint = "^63.0"
sentry-sdk = { version = "^2.17", extras = ["django"] }
argon2-cffi = "^23.1"
gunicorn = "^23.0"
uvicorn = { version = "^0.32", extras = ["standard"] }
httpx = "^0.27"
slack-sdk = "^3.41.0"

[tool.poetry.group.dev.dependencies]
ruff = "^0.7"
mypy = "^1.13"
django-stubs = { version = "^5.1", extras = ["compatible-mypy"] }
djangorestframework-stubs = { version = "^3.15", extras = ["compatible-mypy"] }
pytest = "^8.3"
pytest-django = "^4.9"
pytest-cov = "^5.0"
pytest-mock = "^3.14"
pytest-xdist = "^3.6"
factory-boy = "^3.3"
ipython = "^8.29"
django-debug-toolbar = "^4.4"

[tool.poetry.group.docs]
optional = true

[tool.poetry.group.docs.dependencies]
# Placeholder — add as needed when docs tooling is configured

[build-system]
requires = ["poetry-core>=1.8.0"]
build-backend = "poetry.core.masonry.api"

# ---------------------------------------------------------------------------
# Ruff (lint + format)
# ---------------------------------------------------------------------------
[tool.ruff]
line-length = 100
target-version = "py312"
exclude = [
    ".venv",
    "venv",
    "migrations",
    ".git",
    ".mypy_cache",
    ".ruff_cache",
    "build",
    "dist",
]

[tool.ruff.lint]
select = [
    "E",    # pycodestyle errors
    "W",    # pycodestyle warnings
    "F",    # pyflakes
    "I",    # isort
    "B",    # flake8-bugbear
    "C4",   # flake8-comprehensions
    "UP",   # pyupgrade
    "SIM",  # flake8-simplify
    "RUF",  # ruff-specific
    "DJ",   # flake8-django
]
ignore = [
    "E501",  # line too long (handled by formatter)
    "B008",  # Do not perform function call in argument defaults (common in Django)
]

[tool.ruff.lint.per-file-ignores]
"**/migrations/*.py" = ["E501", "F401"]
"**/settings/*.py" = ["F403", "F405"]

[tool.ruff.format]
quote-style = "double"
indent-style = "space"
line-ending = "auto"

# ---------------------------------------------------------------------------
# mypy
# ---------------------------------------------------------------------------
[tool.mypy]
python_version = "3.12"
strict = true
ignore_missing_imports = true
disable_error_code = ["import-untyped", "misc"]
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true
disallow_incomplete_defs = true
check_untyped_defs = true
no_implicit_optional = true
warn_redundant_casts = true
warn_unused_ignores = true
warn_no_return = true
warn_unreachable = true
plugins = [
    "mypy_django_plugin.main",
    "mypy_drf_plugin.main",
]
exclude = [
    "migrations/",
    "venv/",
    ".venv/",
]

[tool.django-stubs]
django_settings_module = "config.settings.dev"

[[tool.mypy.overrides]]
module = "*.migrations.*"
ignore_errors = true

[[tool.mypy.overrides]]
module = [
    "django_tenants.*",
    "celery.*",
    "environ.*",
    "allauth.*",
    "debug_toolbar.*",
    "anthropic.*",
    "openai.*",
    "pgvector.*",
    "weasyprint.*",
]
ignore_missing_imports = true

# ---------------------------------------------------------------------------
# pytest
# ---------------------------------------------------------------------------
[tool.pytest.ini_options]
DJANGO_SETTINGS_MODULE = "config.settings.test"
python_files = ["test_*.py", "*_test.py", "tests.py"]
python_classes = ["Test*"]
python_functions = ["test_*"]
addopts = [
    "--strict-markers",
    "--tb=short",
    "--reuse-db",
]
markers = [
    "slow: marks tests as slow (deselect with '-m \"not slow\"')",
    "integration: marks tests as integration tests (real external calls)",
    "tenant_isolation: marks tenant isolation verification tests",
]
filterwarnings = [
    "ignore::DeprecationWarning",
    "ignore::PendingDeprecationWarning",
]

[tool.coverage.run]
source = ["apps"]
omit = [
    "*/migrations/*",
    "*/tests/*",
    "*/test_*.py",
    "*/admin.py",
    "*/apps.py",
    "config/settings/*",
    "manage.py",
]

[tool.coverage.report]
exclude_lines = [
    "pragma: no cover",
    "def __repr__",
    "raise NotImplementedError",
    "if TYPE_CHECKING:",
    "if __name__ == .__main__.:",
]
